Don't create orphans when deleting compound (#146)
* Don't create orphans when deleting compound

Instead of just removing relationships from children when deleting a
compound we check if the children have parents. If they do we remove
the relationships, otherwise we delete the children.

* Slay the coding standards dragon.

* Coding standards

Bow down to the tyrany of the coding standards gods! May they take
this scarificial commit and allow my code to pass freely into the
land of review.

* Loosen parent check.